Clean Lua stack between animation cycles
Why just yell if we can fix the problem?  Arguably it might be hiding
bugs, but it might be seen as a feature as well.  If a bug won't
manifest is it really a bug?

I'm coming to the conclusion that it's overly difficult (impractical
or even impossible?) to maintain strict control or ensure correctness
of any sufficiently complex program.

A more relaxed, direct, structure-preserving and locally autonomous
programming paradigm seems to be the most suitable.